Update DB creature_template

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -494,7 +494,7 @@ void ObjectMgr::LoadCreatureTemplate(Field* fields)
     creatureTemplate.unit_flags2            = fields[30].GetUInt32();
     creatureTemplate.unit_flags3            = fields[31].GetUInt32();
     creatureTemplate.dynamicflags           = fields[32].GetUInt32();
-    creatureTemplate.family                 = CreatureFamily(fields[33].GetUInt8());
+    creatureTemplate.family                 = CreatureFamily(fields[33].GetUInt32());
     creatureTemplate.trainer_class          = uint32(fields[34].GetUInt8());
     creatureTemplate.type                   = uint32(fields[35].GetUInt8());
     creatureTemplate.type_flags             = fields[36].GetUInt32();
